requirements.txt
click==6.6
cloudant==2.3.1
ordereddict==1.1
pysolr==3.5.0
python-dotenv==0.5.1
requests==2.11.0
six==1.10.0
slackclient==1.0.1
watson-developer-cloud==0.18.0
websocket-client==0.37.0
